backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.8.2
python-multipart==0.0.9
passlib==1.7.4
PyJWT==2.9.0
sympy==1.13.2
PyYAML==6.0.2
pydantic-settings==2.4.0
# Новые зависимости (v22): карточки учеников с QR-кодом
qrcode[pil]==7.4.2
reportlab==4.2.5
Pillow==10.4.0
